requirements.txt
kafka-python-ng
requests
streamlit
pandas
python-dotenv
watchdog